changed floats into doubles

+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/Bean/DettaglioOrdine.java
@@ -1,5 +1,5 @@
 package it.polimi.tiw.Bean;
 
-public record DettaglioOrdine(int idProdotto, int idFornitore, int idOrdine, int quantita, float prezzo) {
+public record DettaglioOrdine(int idProdotto, int idFornitore, int idOrdine, int quantita, double prezzo) {
 	
 }
