+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/Bean/FasciaDiSpedizione.java
@@ -1,5 +1,5 @@
 package it.polimi.tiw.Bean;
 
-public record FasciaDiSpedizione(int id, int idFornitore, int min, int max, float prezzo ) {
+public record FasciaDiSpedizione(int id, int idFornitore, int min, int max, double prezzo ) {
 	
 }
