+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/Bean/Fornitore.java
@@ -1,6 +1,6 @@
 package it.polimi.tiw.Bean;
 
 
-public record Fornitore(int id, String nome, float valutazione, float soglia) {
+public record Fornitore(int id, String nome, double valutazione, double soglia) {
 	
 }
