+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/Bean/Ordine.java
@@ -1,5 +1,5 @@
 package it.polimi.tiw.Bean;
 
-public record Ordine(int id, float tot, float spedizione, String indirizzo, String email) {
+public record Ordine(int id, double tot, double spedizione, String indirizzo, String email) {
 	
 }
